restructure shi
